Extract Formatter etc. into formatting.py; mv pars{er,ing}.py
